@@ -4,7 +4,8 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --aot --proxy-config proxy.config.json",
+		"start": "ng serve --aot --proxy-config proxy.config.json",
+		"start:fr": "ng serve --aot --i18nFile=src/locale/messages.fr.xlf --i18nFormat=xlf --locale=fr --proxy-config proxy.config.json",
     "build": "ng build --aot --prod --build-optimizer",
     "test": "ng test",
     "lint": "ng lint",
@@ -12,17 +13,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.0",
+    "@angular/animations": "^5.2.0",
     "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "^5.1.0",
-    "@angular/compiler": "^5.1.0",
-    "@angular/core": "^5.1.0",
-    "@angular/forms": "^5.1.0",
-    "@angular/http": "^5.1.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
     "@angular/material": "^5.0.0",
-    "@angular/platform-browser": "^5.1.0",
-    "@angular/platform-browser-dynamic": "^5.1.0",
-    "@angular/router": "^5.1.0",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/router": "^5.2.0",
     "@ngrx/core": "^1.2.0",
     "@ngrx/effects": "^4.1.0",
     "@ngrx/router-store": "^4.1.0",
@@ -33,10 +34,7 @@
     "angular-file-dropzone": "^2.1.1",
     "core-js": "^2.4.1",
     "d3": "^4.11.0",
-    "dragula": "^3.7.2",
     "font-awesome": "^4.7.0",
-    "ng2-dnd": "^5.0.2",
-    "ng2-dragula": "^1.5.0",
     "ngrx-store-freeze": "^0.2.0",
     "ngrx-store-logger": "^0.2.0",
     "reselect": "^3.0.1",
@@ -45,7 +43,7 @@
   },
   "devDependencies": {
     "@angular-devkit/schematics": "0.0.40",
-    "@angular/cli": "^1.6.1",
+    "@angular/cli": "^1.6.3",
     "@angular/compiler-cli": "^5.1.0",
     "@angular/language-service": "^5.1.0",
     "@types/jasmine": "~2.5.53",
