@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { KanbanDirective } from './directives/kanban.directive';
 import { KanbanColComponent } from './components/kanban-col/kanban-col.component';
 import { KanbanItemComponent } from './components/kanban-item/kanban-item.component';
 import { KanbanService } from './services/kanban.service';
@@ -9,7 +8,7 @@ import { KanbanService } from './services/kanban.service';
 	imports: [
 		CommonModule
 	],
-	declarations: [KanbanDirective, KanbanColComponent, KanbanItemComponent],
+	declarations: [ KanbanColComponent, KanbanItemComponent],
 	exports: [ KanbanColComponent, KanbanItemComponent ],
 	providers: [ KanbanService ]
 })
