@@ -3,7 +3,7 @@ import { ChangeDetectorRef } from '@angular/core';
 import { SelectableItem } from '../selectable-item.interface';
 import { AbstractInput, makeAccessorProvider } from '../../../inputs/abstract-input.class';
 import Log from '../../../../../utils/logger/log.class';
-import { ChangeDetectionStrategy } from '@angular/core/src/change_detection/constants';
+import { ChangeDetectionStrategy } from '@angular/core';
 
 @Component({
 	selector: 'input-select-multi-app',
